refactor(jans-cedarling): move id token trust mode to authz config
