+++ jans-cedarling/cedarling/src/bootstrap_config/decode.rs
@@ -3,21 +3,19 @@
 //
 // Copyright (c) 2024, Gluu, Inc.
 
+use super::authorization_config::{AuthorizationConfig, IdTokenTrustMode};
+use super::{
+    BootstrapConfig, BootstrapConfigLoadingError, JwtConfig, LogConfig, LogTypeConfig,
+    MemoryLogConfig, PolicyStoreConfig, PolicyStoreSource, TokenValidationConfig,
+};
+use crate::log::LogLevel;
+use jsonwebtoken::Algorithm;
+use serde::{Deserialize, Deserializer, Serialize};
 use std::collections::HashSet;
 use std::fmt::Display;
 use std::fs;
 use std::path::Path;
 use std::str::FromStr;
-
-use jsonwebtoken::Algorithm;
-use serde::{Deserialize, Deserializer, Serialize};
-
-use super::authorization_config::AuthorizationConfig;
-use super::{
-    BootstrapConfig, BootstrapConfigLoadingError, IdTokenTrustMode, JwtConfig, LogConfig,
-    LogTypeConfig, MemoryLogConfig, PolicyStoreConfig, PolicyStoreSource, TokenValidationConfig,
-};
-use crate::log::LogLevel;
 
 #[derive(Deserialize, PartialEq, Debug, Default)]
 /// Struct that represent mapping mapping `Bootstrap properties` to be JSON and YAML compatible
@@ -503,7 +501,6 @@ impl BootstrapConfig {
             jwks,
             jwt_sig_validation: raw.jwt_sig_validation.into(),
             jwt_status_validation: raw.jwt_status_validation.into(),
-            id_token_trust_mode: raw.id_token_trust_mode,
             signature_algorithms_supported: raw.jwt_signature_algorithms_supported.clone(),
             access_token_config: TokenValidationConfig {
                 iss_validation: raw.at_iss_validation.into(),
@@ -541,6 +538,7 @@ impl BootstrapConfig {
             mapping_id_token: raw.mapping_id_token.clone(),
             mapping_access_token: raw.mapping_access_token.clone(),
             mapping_userinfo_token: raw.mapping_userinfo_token.clone(),
+            id_token_trust_mode: raw.id_token_trust_mode,
         };
 
         Ok(Self {
