+++ jans-cedarling/cedarling/src/bootstrap_config/jwt_config.rs
@@ -3,11 +3,8 @@
 //
 // Copyright (c) 2024, Gluu, Inc.
 
-use std::collections::HashSet;
-use std::str::FromStr;
-
 use jsonwebtoken::Algorithm;
-use serde::Deserialize;
+use std::collections::HashSet;
 
 /// The set of Bootstrap properties related to JWT validation.
 #[derive(Debug, PartialEq)]
@@ -38,23 +35,6 @@ pub struct JwtConfig {
     ///
     /// [`IETF Draft`]: https://datatracker.ietf.org/doc/draft-ietf-oauth-status-list/
     pub jwt_status_validation: bool,
-    /// Sets the validation level for ID tokens.
-    ///
-    /// The available levels are [`None`] and [`Strict`].
-    ///
-    /// # Strict Mode
-    ///
-    /// In `Strict` mode, the following conditions must be met for a token
-    /// to be considered valid:
-    ///
-    /// - The `id_token`'s `aud` (audience) must match the `access_token`'s `client_id`
-    /// - If a Userinfo token is present:
-    ///     - Its `sub` (subject) must match the `id_token`'s `sub`.
-    ///     - Its `aud` (audience) must match the `access_token`'s `client_id`.
-    ///
-    /// [`None`]: IdTokenTrustMode::None
-    /// [`Strict`]: IdTokenTrustMode::Strict
-    pub id_token_trust_mode: IdTokenTrustMode,
     /// Only tokens signed with algorithms in this list can be valid.
     pub signature_algorithms_supported: HashSet<Algorithm>,
     /// Validation options related to the Access token
@@ -177,44 +157,6 @@ impl TokenValidationConfig {
     }
 }
 
-/// Defines the level of validation for ID tokens.
-#[derive(Debug, Clone, PartialEq, Default, Deserialize, Copy)]
-#[serde(rename_all = "lowercase")]
-pub enum IdTokenTrustMode {
-    /// No validation is performed on the ID token.
-    None,
-    /// Strict validation of the ID token.
-    ///
-    /// In this mode, the following conditions must be met:
-    ///
-    /// - The `id_token`'s `aud` (audience) must match the `access_token`'s `client_id`.
-    /// - If a Userinfo token is present:
-    ///   - Its `sub` (subject) must match the `id_token`'s `sub`.
-    ///   - Its `aud` must match the `access_token`'s `client_id`.
-    #[default]
-    Strict,
-}
-
-impl FromStr for IdTokenTrustMode {
-    type Err = IdTknTrustModeParseError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let s = s.to_lowercase();
-        match s.as_str() {
-            "strict" => Ok(IdTokenTrustMode::Strict),
-            "none" => Ok(IdTokenTrustMode::None),
-            _ => Err(IdTknTrustModeParseError { trust_mode: s }),
-        }
-    }
-}
-
-/// Error when parsing [`IdTokenTrustMode`]
-#[derive(Default, Debug, derive_more::Display, derive_more::Error)]
-#[display("Invalid `IdTokenTrustMode`: {trust_mode}. should be `strict` or `none`")]
-pub struct IdTknTrustModeParseError {
-    trust_mode: String,
-}
-
 impl Default for JwtConfig {
     /// Cedarling will use the strictest validation options by default.
     fn default() -> Self {
@@ -222,7 +164,6 @@ impl Default for JwtConfig {
             jwks: None,
             jwt_sig_validation: true,
             jwt_status_validation: true,
-            id_token_trust_mode: IdTokenTrustMode::Strict,
             signature_algorithms_supported: HashSet::new(),
             access_token_config: TokenValidationConfig::access_token(),
             id_token_config: TokenValidationConfig::id_token(),
@@ -238,7 +179,6 @@ impl JwtConfig {
             jwks: None,
             jwt_sig_validation: false,
             jwt_status_validation: false,
-            id_token_trust_mode: IdTokenTrustMode::None,
             signature_algorithms_supported: HashSet::new(),
             access_token_config: TokenValidationConfig::default(),
             id_token_config: TokenValidationConfig::default(),
